merged with latest status from badaix

+++ common/message/client_info.hpp
@@ -40,7 +40,7 @@ public:
 
     uint16_t getVolume()
     {
-        return get("volume", 100);
+        return get("volume", static_cast<uint16_t>(100));
     }
 
     bool isMuted()
@@ -58,7 +58,7 @@ public:
         msg["muted"] = muted;
     }
 };
-}
+} // namespace msg
 
 
 #endif
